Fixed multiple calls to a filter
